Feature/#491/INTERVIEW CLIENT - PUT VIDEO CHAT LOGIC IN CONTEXT (#492)
* FEAT/Placed all stream logic in a Context

* FIX/Added some comments
